requirements.txt
openai==1.65.1
langchain==0.3.20
langchain-openai==0.2.11
streamlit==1.44.1
python-dotenv==1.0.1
pypdf==4.2.0
python-docx==1.1.2
faiss-cpu==1.8.0
tiktoken==0.7.0
